Reusable logic pivot and heatmap + update imports
