+++ src/components/charts/tables/tables.hooks.ts
@@ -1,0 +1,134 @@
+import { DataResponse, Dimension, Measure } from '@embeddable.com/core';
+import { useMemo } from 'react';
+
+type Order = 'asc' | 'desc' | 'equal';
+
+function getTablePredominantOrder(arr: (string | number | boolean)[]): Order {
+  let asc = 0;
+  let desc = 0;
+
+  for (let i = 0; i < arr.length - 1; i++) {
+    const a = arr[i];
+    const b = arr[i + 1];
+
+    if (a == null || b == null) continue;
+
+    if (a < b) asc++;
+    else if (a > b) desc++;
+  }
+
+  if (asc > desc) return 'asc';
+  if (desc > asc) return 'desc';
+  return 'equal';
+}
+
+function isTableOrderMixed(arr: (string | number | boolean)[]): boolean {
+  let asc = 0;
+  let desc = 0;
+
+  for (let i = 0; i < arr.length - 1; i++) {
+    const a = arr[i];
+    const b = arr[i + 1];
+
+    if (a == null || b == null) continue;
+
+    if (a < b) asc++;
+    else if (a > b) desc++;
+  }
+
+  return asc > 0 && desc > 0;
+}
+
+const getTableSortComparator = (order: Order) => {
+  return <T>(a: T, b: T) => {
+    if (order === 'asc') {
+      return a < b ? -1 : a > b ? 1 : 0;
+    }
+    if (order === 'desc') {
+      return a > b ? -1 : a < b ? 1 : 0;
+    }
+    return 0;
+  };
+};
+
+type useGetTableSortedResultsProps = {
+  results: DataResponse;
+  columnOrder: (string | number | boolean)[];
+  rowOrder: (string | number | boolean)[];
+  columnDimension: Dimension;
+  rowDimension: Dimension;
+  measures: Measure[];
+};
+
+export const getTableSortedResults = (props: useGetTableSortedResultsProps) => {
+  const data = props.results.data ?? [];
+  const firstRow = data[0];
+
+  if (!firstRow) return [];
+
+  const buildEmptyMeasures = () =>
+    props.measures.reduce<Record<string, unknown>>((acc, measure) => {
+      acc[measure.name] = undefined;
+      return acc;
+    }, {});
+
+  const reorderByAxis = (
+    axisOrder: (string | number | boolean)[],
+    axisDimension: Dimension,
+    fixedDimension: Dimension,
+  ) => {
+    if (!isTableOrderMixed(axisOrder)) {
+      return data;
+    }
+
+    const order = getTablePredominantOrder(axisOrder);
+    const sortedAxisOrder = [...axisOrder].sort(getTableSortComparator(order));
+
+    // “Leader” rows: one per value in the axis order, at the fixed dimension’s first value
+    const leaderRows = sortedAxisOrder.map((value) => {
+      const existing = data.find(
+        (x) =>
+          x[axisDimension.name] === value &&
+          x[fixedDimension.name] === firstRow[fixedDimension.name],
+      );
+
+      if (existing) return existing;
+
+      return {
+        ...firstRow,
+        [axisDimension.name]: value,
+        ...buildEmptyMeasures(),
+      };
+    });
+
+    // Remove rows that are already covered by leaderRows (same [axis, fixed] tuple)
+    const restResults = data.filter(
+      (resultRow) =>
+        !leaderRows.some(
+          (leaderRow) =>
+            leaderRow[axisDimension.name] === resultRow[axisDimension.name] &&
+            leaderRow[fixedDimension.name] === resultRow[fixedDimension.name],
+        ),
+    );
+
+    return [...leaderRows, ...restResults];
+  };
+
+  // Preserve original behavior: fix columns if mixed, otherwise rows if mixed
+  if (isTableOrderMixed(props.columnOrder)) {
+    return reorderByAxis(props.columnOrder, props.columnDimension, props.rowDimension);
+  }
+
+  if (isTableOrderMixed(props.rowOrder)) {
+    return reorderByAxis(props.rowOrder, props.rowDimension, props.columnDimension);
+  }
+
+  return data;
+};
+
+export const useGetTableSortedResults = (props: useGetTableSortedResultsProps) => {
+  const results = useMemo(() => {
+    return getTableSortedResults(props);
+  }, [props]);
+  return results;
+};
